Move TransmutationsBook to root package
