+++ Alchemist/src/main/java/com/brunodles/alchemist/MethodToAnnotationInvocationHandler.java
@@ -3,7 +3,6 @@ package com.brunodles.alchemist;
 import com.brunodles.alchemist.exceptions.ResultException;
 import com.brunodles.alchemist.exceptions.TransformerException;
 import com.brunodles.alchemist.selector.MissingSelectorException;
-import com.brunodles.alchemist.transformers.TransmutationsBook;
 import org.jetbrains.annotations.NotNull;
 
 import java.lang.annotation.Annotation;
