+++ Alchemist/src/test/java/com/brunodles/alchemist/AlchemistTest/WhenLastResultIsEmpty.java
@@ -3,8 +3,7 @@ package com.brunodles.alchemist.AlchemistTest;
 import com.brunodles.alchemist.Alchemist;
 import com.brunodles.alchemist.AnnotationInvocation;
 import com.brunodles.alchemist.AnnotationTransmutation;
-import com.brunodles.alchemist.Transmutation;
-import com.brunodles.alchemist.transformers.TransmutationsBook;
+import com.brunodles.alchemist.TransmutationsBook;
 import org.junit.Assert;
 import org.junit.Test;
 import org.junit.runner.RunWith;
